Make it possibel to update vehicle by clicking on it
