+++ app/build.gradle
@@ -52,7 +52,7 @@ android {
     }
 }
 
-def BUTTER_KNIFE_VERSION="8.8.1"
+def BUTTER_KNIFE_VERSION = "8.8.1"
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
@@ -69,7 +69,8 @@ dependencies {
 
     // Dagger 2
     implementation 'com.google.dagger:dagger-android:2.15'
-    implementation 'com.google.dagger:dagger-android-support:2.15' // if you use the support libraries
+    implementation 'com.google.dagger:dagger-android-support:2.15'
+    // if you use the support libraries
     annotationProcessor 'com.google.dagger:dagger-android-processor:2.15'
     annotationProcessor 'com.google.dagger:dagger-compiler:2.15'
